ACKNOWLEDGED - # 1924: Fix Early Censored Run Handling https://mantis.sjrx.net/view.php?id=1924
A bunch of changes in this build.

1) Added a new isCensoredEarly() method AlgorithmRun as the logic seemed
to be strewn accross many classes and done inconsistently.

2) Removed two stale test cases related to functional anova.

3) Refactored the RunHistory interface to make it cleaner and easier to
deal with redundant runs. This includes:
   A) Explicitly marking all methods with whether they Include or
Exclude Redundant runs (a run is redundant if we have a better lower
bound for it's perfomance). 
   B) Replaced many occurances of getting all runs, and then size() with
the actual faster method of just computing the the direct time.
   C) Moved the getCensoredFlag and getRunResponse values out of the
interface and into RunHistoryHelper, this lessens the chance that the
values of the methods will be in the wrong order, because of new runs
appearing, or weirdo iteration order. Additionally it's probably faster
since internally those methods called the same method which rebuilds the
list every time.